fix(review): move recipe envelope construction into delivery domain
