+++ src/autoskillit/server/tools/_serve_helpers.py
@@ -16,18 +16,11 @@ from typing import TYPE_CHECKING, Any
 from autoskillit.core import (
     RESPONSE_BACKSTOP_EXEMPTION_REGISTRY,
     RESPONSE_BACKSTOP_EXEMPTION_REGISTRY_DIGEST,
-    get_logger,
-)
-from autoskillit.recipe import (  # noqa: F401 — canonical extractor reused by build_routing_edges_by_step default
-    _extract_routing_edges,
-    step_byte_ranges_from_yaml,
 )
 
 if TYPE_CHECKING:
     from autoskillit.core import BackendCapabilities, CodingAgentBackend
     from autoskillit.pipeline.context import ToolContext
-    from autoskillit.recipe.schema import RecipeStep
-    from autoskillit.server._recipe_delivery import RecipeArtifactGeneration
 
 
 def build_backend_capabilities_map(
@@ -192,399 +185,3 @@ def serve_recipe(
     if ctx.recipes is None:
         raise RuntimeError("serve_recipe() called with ctx.recipes=None")
     return ctx.recipes.load_and_validate(name, ctx.project_dir, **kwargs)
-
-
-# ---------------------------------------------------------------------------
-# Bounded envelope (Part B of #4304)
-# ---------------------------------------------------------------------------
-
-
-def _step_one_line_summary(step: RecipeStep) -> str:
-    """Return a compact one-line summary of *step* for the envelope skeleton.
-
-    Used by the open_kitchen / load_recipe integration paths to populate
-    ``step_summaries`` before calling ``extract_step_skeleton``. Prefers
-    an explicit ``description`` field; falls back to a 160-char head of
-    the step's ``message`` or a tool/action signature. Always single-line.
-    """
-    desc = getattr(step, "description", "") or ""
-    if desc.strip():
-        return desc.strip().splitlines()[0][:160]
-    msg = getattr(step, "message", None)
-    if isinstance(msg, str) and msg.strip():
-        return msg.strip().splitlines()[0][:160]
-    tool = getattr(step, "tool", None)
-    action = getattr(step, "action", None)
-    if tool:
-        return f"tool={tool}"
-    if action:
-        return f"action={action}"
-    return ""
-
-
-def _compute_step_byte_ranges(content: str) -> dict[str, tuple[int, int]]:
-    """Return ``{step_name: (start, end)}`` UTF-8 byte offsets within *content*.
-
-    Thin wrapper over
-    :func:`autoskillit.recipe.step_byte_ranges_from_yaml` that keeps the
-    ``_serve_helpers``-local symbol stable for downstream callers and
-    tests. The yaml-handling and ``isinstance(..., MappingNode)``
-    fail-open logic lives in ``core.io`` so that yaml imports are
-    confined to the single canonical module (the package-wide invariant
-    enforced by ``tests/arch/test_subpackage_isolation.py`` and
-    ``tests/core/test_io.py::TestYamlConsolidationArchitecture``).
-    """
-    return step_byte_ranges_from_yaml(content)
-
-
-def extract_step_skeleton(
-    post_prune_step_names: list[str],
-    routing_edges_by_step: dict[str, list[tuple[str, str]]],
-    step_summaries: dict[str, str] | None = None,
-    byte_ranges: dict[str, tuple[int, int]] | None = None,
-) -> dict[str, Any]:
-    """Build a compact step-flow skeleton from parsed post-prune step data.
-
-    The skeleton is a list of per-step dicts (name + one-line summary +
-    outgoing routing edges + optional byte range). Combined with the
-    byte-range index in the persisted artifact, the orchestrator can
-    route without pulling a full step body, and pull-on-demand only the
-    step it is about to execute.
-
-    *post_prune_step_names* — order-preserving list of step names (from
-        ``load_and_validate``'s ``post_prune_step_names`` result field).
-    *routing_edges_by_step* — name → list of (edge_type, target) tuples
-        derived from ``_extract_routing_edges`` for each step.
-    *step_summaries* — optional name → one-line summary override. A step
-        with no summary (or an empty one) omits the ``summary`` key
-        entirely rather than emitting it as ``""`` — saves bytes at scale
-        and matches the byte_range fail-open/omit convention below.
-    *byte_ranges* — optional name → (start, end) UTF-8 byte offsets
-        within the persisted ``content`` field. When provided for a
-        given step, the skeleton's per-step entry gains a
-        ``byte_range`` field carrying ``[start, end]``; steps absent
-        from the map get no such key (fail-open, matching the
-        edges/summary fallback pattern above).
-    """
-    skeleton: list[dict[str, Any]] = []
-    for name in post_prune_step_names:
-        edges = routing_edges_by_step.get(name) or []
-        summary = (step_summaries or {}).get(name) or ""
-        entry: dict[str, Any] = {
-            "name": name,
-            "edges": [
-                {"type": edge_type, "target": target} for edge_type, target in edges if target
-            ],
-        }
-        if summary:
-            entry["summary"] = summary
-        span = (byte_ranges or {}).get(name)
-        if span is not None:
-            entry["byte_range"] = list(span)
-        skeleton.append(entry)
-    return {
-        "step_count": len(skeleton),
-        "steps": skeleton,
-    }
-
-
-def build_recipe_envelope(
-    payload: dict[str, Any],
-    *,
-    recipe_name: str,
-    generation: RecipeArtifactGeneration,
-    skeleton_source: ToolContext,
-    bound_bytes: int,
-) -> dict[str, Any]:
-    """Build a bounded envelope that fits the smallest backend delivery bound.
-
-    The envelope is the orchestrator-visible replacement for the full
-    recipe payload when the payload exceeds a backend's effective delivery
-    token limit. It carries:
-
-    - routing metadata (``success``, ``kitchen``, ``version``, ``valid``,
-      ``dispatch_feasible``);
-    - verbatim priority content (``orchestration_rules``,
-      ``stop_step_semantics``, ``errors``, ``warnings``, ``hooks``);
-    - ingredients schema (``ingredients_table``, ``suggestions``);
-    - the step-flow skeleton (post-prune step names, summaries, routing
-      edges) — enough for the orchestrator to route between steps without
-      pulling a full step body;
-    - the post-prune step list and routing edge list (so callers that
-      only need the routing graph don't need to parse the skeleton);
-    - a pull reference pointing to the deterministic artifact path
-      (overwritten by every open_kitchen / load_recipe call) and the
-      ``get_recipe_section`` MCP tool name.
-
-    The envelope omits the full ``content`` field. The orchestrator pulls
-    each step's body via ``get_recipe_section(section=<step_name>)`` at
-    execution time, bounded to the delivery limit and chunked with a
-    continuation token for steps larger than one chunk.
-
-    ``bound_bytes`` is the effective delivery byte ceiling (smallest
-    backend bound × 4). Large string fields like ``orchestration_rules``
-    are projected to fit alongside the skeleton + pull reference via a
-    two-phase allocator: phase 1 drops (omits, never emits as ``""``)
-    any key that can't even afford its JSON-key overhead under the
-    current ``remaining`` budget; phase 2 water-fills the leftover
-    content budget across survivors so neither priority field can
-    monopolize the pool. Truncation goes through ``_safe_utf8_truncate``
-    so a multi-byte UTF-8 codepoint is never split into an
-    undecodable byte sequence. ``recipe_name`` is required (keyword-only)
-    to match the sibling-function convention in this module.
-    """
-    post_prune_raw = payload.get("post_prune_step_names") or []
-    if not isinstance(post_prune_raw, list):
-        post_prune_raw = []
-    post_prune_names = [name for name in post_prune_raw if isinstance(name, str)]
-    active_recipe_steps: dict[str, Any] | None = None
-    if getattr(skeleton_source, "recipe_name", "") == recipe_name:
-        active_recipe_steps = skeleton_source.active_recipe_steps
-    summaries = build_step_summaries(active_recipe_steps)
-    edges = build_routing_edges_by_step(active_recipe_steps)
-    byte_ranges = _compute_step_byte_ranges(payload.get("content") or "")
-    skeleton = extract_step_skeleton(
-        post_prune_names,
-        edges,
-        summaries,
-        byte_ranges=byte_ranges,
-    )
-    pull_identity = generation.pull_identity()
-
-    def _pullable_skeleton_size(candidate: dict[str, Any]) -> int:
-        pullable = {
-            "success": payload.get("success", True),
-            "step_flow_skeleton": candidate,
-            "recipe_pull": pull_identity,
-            "delivery_bound_spill": True,
-        }
-        return len(json.dumps(pullable, ensure_ascii=False, separators=(",", ":")).encode("utf-8"))
-
-    if _pullable_skeleton_size(skeleton) > bound_bytes:
-        for summary_limit in (120, 80, 64, 48, 32, 24, 16, 8, 0):
-            bounded_summaries = (
-                {name: summary[:summary_limit] for name, summary in summaries.items()}
-                if summary_limit
-                else {}
-            )
-            skeleton = extract_step_skeleton(
-                post_prune_names,
-                edges,
-                bounded_summaries,
-                byte_ranges=byte_ranges,
-            )
-            if _pullable_skeleton_size(skeleton) <= bound_bytes:
-                break
-
-    envelope: dict[str, Any] = {}
-    envelope_bytes = 0
-    serialized = json.dumps(payload.get("success", True)).encode("utf-8")
-    envelope_bytes += len(serialized)
-    envelope["success"] = payload.get("success", True)
-    for key in (
-        "kitchen",
-        "version",
-        "valid",
-        "dispatch_feasible",
-        "errors",
-        "warnings",
-        "hooks",
-        "post_prune_step_names",
-        "post_prune_routing_edges",
-        "requires_packs",
-        "requires_features",
-    ):
-        if key in payload and payload[key] is not None:
-            envelope[key] = payload[key]
-            envelope_bytes += len(
-                json.dumps({key: payload[key]}, ensure_ascii=False).encode("utf-8")
-            )
-
-    # Fixed-size overhead: skeleton JSON + pull reference + delivery_bound_spill.
-    skeleton_overhead = len(
-        json.dumps({"step_flow_skeleton": skeleton}, ensure_ascii=False).encode("utf-8")
-    )
-    pull_overhead = len(
-        json.dumps(
-            {
-                "recipe_pull": pull_identity,
-                "delivery_bound_spill": True,
-            },
-            ensure_ascii=False,
-        ).encode("utf-8")
-    )
-
-    remaining = max(
-        0,
-        bound_bytes - skeleton_overhead - pull_overhead - envelope_bytes - 64,
-    )
-
-    def _project_priority_strings(keys: tuple[str, ...]) -> None:
-        nonlocal remaining
-        candidates = [k for k in keys if isinstance(payload.get(k), str) and payload.get(k)]
-        if not candidates:
-            return
-
-        overhead = {
-            k: len(json.dumps({k: ""}, ensure_ascii=False).encode("utf-8")) for k in candidates
-        }
-        # Phase 1: a key survives only if its JSON-key overhead plus at least
-        # one content byte fits what's left, checked in priority (declaration)
-        # order. Keys that don't clear this bar are omitted entirely — never
-        # emitted as "".
-        present: list[str] = []
-        budget = remaining
-        for k in candidates:
-            if overhead[k] < budget:
-                present.append(k)
-                budget -= overhead[k]
-        if not present:
-            return
-
-        # Phase 2: water-fill the leftover content budget evenly across
-        # survivors, round by round. A key whose value is shorter than its
-        # share drops out and its unused share carries over to the *other*
-        # survivors in the next round — no single key can monopolize the
-        # pool the way a sequential first-come-first-served allocator can.
-        content_budget = remaining - sum(overhead[k] for k in present)
-        lengths = {k: len(payload[k].encode("utf-8")) for k in present}
-        alloc: dict[str, int] = dict.fromkeys(present, 0)
-        active = list(present)
-        pool = content_budget
-        while active and pool > 0:
-            share, extra = divmod(pool, len(active))
-            still_active: list[str] = []
-            for index, key in enumerate(active):
-                give = share + (1 if index < extra else 0)
-                need = lengths[key] - alloc[key]
-                take = min(give, need, pool)
-                alloc[key] += take
-                pool -= take
-                if alloc[key] < lengths[key]:
-                    still_active.append(key)
-            active = still_active
-
-        for key in present:
-            take = alloc[key]
-            if take <= 0:
-                continue  # no content budget survives for this field; omit it
-            value_bytes = payload[key].encode("utf-8")
-            if len(value_bytes) <= take:
-                envelope[key] = payload[key]
-                remaining -= len(value_bytes) + overhead[key]
-            else:
-                envelope[key] = _safe_utf8_truncate(value_bytes[:take])
-                remaining -= take + overhead[key]
-                get_logger(__name__).warning(
-                    "recipe_envelope_priority_field_truncated",
-                    recipe_name=recipe_name,
-                    field=key,
-                    alloc_bytes=take,
-                )
-
-    _project_priority_strings(("orchestration_rules", "stop_step_semantics"))
-
-    # ingredients_table and suggestions are deprioritized — serialize them
-    # only if budget allows; otherwise omit. The orchestrator can pull the
-    # full ingredients_table via ``get_recipe_section(section="ingredients_table")``.
-    for key in ("ingredients_table", "suggestions"):
-        value = payload.get(key)
-        if value is None:
-            continue
-        serialized_value = json.dumps(value, ensure_ascii=False).encode("utf-8")
-        if len(serialized_value) + 32 <= remaining:
-            envelope[key] = value
-            remaining -= len(serialized_value) + 32
-
-    envelope["step_flow_skeleton"] = skeleton
-    envelope["recipe_pull"] = pull_identity
-    envelope["delivery_bound_spill"] = True
-    if (
-        len(json.dumps(envelope, ensure_ascii=False, separators=(",", ":")).encode("utf-8"))
-        <= bound_bytes
-    ):
-        return envelope
-
-    fallback_candidates: tuple[dict[str, Any], ...] = (
-        {
-            "success": payload.get("success", True),
-            "step_flow_skeleton": skeleton,
-            "recipe_pull": pull_identity,
-            "delivery_bound_spill": True,
-        },
-        {
-            "success": False,
-            "error": "recipe_envelope_exceeds_delivery_bound",
-            "recipe_pull": pull_identity,
-        },
-        {"success": False, "error": "recipe_envelope_exceeds_delivery_bound"},
-        {},
-    )
-    for fallback in fallback_candidates:
-        if (
-            len(json.dumps(fallback, ensure_ascii=False, separators=(",", ":")).encode("utf-8"))
-            <= bound_bytes
-        ):
-            return fallback
-    raise ValueError("delivery bound is too small for a JSON object")
-
-
-def _safe_utf8_truncate(data: bytes) -> str:
-    """Decode *data* as UTF-8, backing off byte-by-byte from the end on failure.
-
-    Handles both a dangling continuation byte and a dangling multi-byte lead
-    byte — the two ways a naive byte-index slice can land mid-codepoint.
-    Used by the envelope priority-field allocator to guarantee that a
-    multi-byte codepoint is never split into an undecodable byte sequence.
-    """
-    while data:
-        try:
-            return data.decode("utf-8")
-        except UnicodeDecodeError as exc:
-            data = data[: exc.start]
-    return ""
-
-
-def build_step_summaries(active_recipe_steps: Any) -> dict[str, str]:
-    """Build a {step_name: one_line_summary} dict from the parsed Recipe.steps.
-
-    Falls back to an empty string per step if the parsed structure is
-    unavailable (e.g. the recipe was loaded but post-prune filtering
-    stripped the steps out, or the serve path passed no Recipe object).
-    """
-    if not isinstance(active_recipe_steps, dict) or not active_recipe_steps:
-        return {}
-    summaries: dict[str, str] = {}
-    for name, step in active_recipe_steps.items():
-        if not isinstance(name, str) or not name:
-            continue
-        summaries[name] = _step_one_line_summary(step)
-    return summaries
-
-
-def build_routing_edges_by_step(
-    active_recipe_steps: Any,
-    *,
-    edge_extractor: Any = _extract_routing_edges,
-) -> dict[str, list[tuple[str, str]]]:
-    """Build a {step_name: [(edge_type, target), ...]} dict via edge_extractor.
-
-    ``edge_extractor`` is the existing ``_extract_routing_edges`` callable
-    from ``recipe/_analysis_graph.py``. We avoid importing it here to keep
-    the helper importable from lightweight contexts; the caller passes the
-    callable in. Steps whose extractor returns nothing map to an empty
-    list (not omitted) so callers don't have to ``.get(name) or []``.
-    """
-    if not isinstance(active_recipe_steps, dict) or not active_recipe_steps:
-        return {}
-    edges_by_step: dict[str, list[tuple[str, str]]] = {}
-    for name, step in active_recipe_steps.items():
-        if not isinstance(name, str) or not name:
-            continue
-        extracted = edge_extractor(step) if edge_extractor is not None else []
-        edges_by_step[name] = [
-            (edge.edge_type, edge.target)
-            for edge in (extracted or [])
-            if getattr(edge, "target", None)
-        ]
-    return edges_by_step
